refactor: rename `_rust` -> `_pure`
