+++ src/encodings/webp_pure.rs
